Update deploy-availability-zone-powershell.md
The behavior of New-AzPublicIpAddress was changed like a following:

Upcoming breaking changes in the cmdlet 'New-AzPublicIpAddress' :
Default behaviour of Zone will be changed
Cmdlet invocation changes :
    Old Way : Sku = Standard means the Standard Public IP is zone-redundant.
    New Way : Sku = Standard and Zone = {} means the Standard Public IP has no zones. If you want to create a zone-redundant Public IP address, please specify all the zones in the region. For example, Zone = ['1', '2', '3'].
Note : Go to https://aka.ms/azps-changewarnings for steps to suppress this breaking change warning, and other information on breaking changes in Azure PowerShell.

To create new Azure Firewall with Zone redundant, it needs to create Public IP address as Zone redundant. This means -Zone parameter should be specified as 1,2,3 in New-AzPublicIpAddress command.